fix: ModalExample focus state ref: Modal example into its own component
